[SQL] Simplify the structure of the simple end to end tests
