+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/simple/OptimizedIncrementalTests.java
@@ -32,8 +32,4 @@ public class OptimizedIncrementalTests extends NaiveIncrementalTests {
         CompilerOptions options = this.testOptions(true, true, false);
         return new DBSPCompiler(options);
     }
-
-    public void invokeTestQueryBase(String query, InputOutputPair... streams) {
-        super.testQueryBase(query, streams);
-    }
 }
